Creating dummy:regenerate Rake task
There is a bit more going on in that I'm removing some of the engine's
default tasks (i.e. those normally namespaced to :app) and instead
providing a bare bones number of rake tasks for interacting with the
engine's dummy application.

I am going to be moving towards having `rake spec` regenerate the entire
dummy application with each invocation. However there are pathing issues
that are occuring.